reorganize units
