+++ exposan/htl/units_to_move/__init__.py
@@ -14,57 +14,52 @@ Please refer to https://github.com/QSD-Group/EXPOsan/blob/main/LICENSE.txt
 for license details.
 '''
 
+# Used in other units, need to be imported first
 from . import (
-    _chg,
-    _chp,    
     _compressor,
+    _heat_exchanging,
+    _pumping,
+    _reactor,
+    )
+from ._compressor import *
+from ._heat_exchanging import *
+from ._pumping import *
+from ._reactor import *
+
+from . import (
+    _abstract,
+    _combustion,
     _distillation,
     _flash,
-    _hc,
-    _ht,
-    _htl,
-    _hx_utility,
+    _hydroprocessing,
+    _hydrothermal,
     _membrane_distillation,
-    _phase_changer,
-    _pump,
-    _reactor,
-    _sludge_centrifuge,
-    _storage_tank,
-    _struvite_precipitation,
+    _sludge_thickening,
+    _tanks,
     )
 
-from ._chg import *
-from ._chp import *
-from ._compressor import *
+from ._abstract import *
+from ._combustion import *
 from ._distillation import *
 from ._flash import *
-from ._hc import *
-from ._ht import *
-from ._htl import *
-from ._hx_utility import *
+from ._hydroprocessing import *
+from ._hydrothermal import *
 from ._membrane_distillation import *
-from ._phase_changer import *
-from ._pump import *
-from ._reactor import *
-from ._sludge_centrifuge import *
-from ._storage_tank import *
-from ._struvite_precipitation import *
+from ._sludge_thickening import *
+from ._tanks import *
 
 __all__ = (
-    *_chg.__all__,
-    *_chp.__all__,
+    *_abstract.__all__,
+    *_combustion.__all__,
     *_compressor.__all__,
     *_distillation.__all__,
     *_flash.__all__,
-    *_hc.__all__,
-    *_ht.__all__,
-    *_htl.__all__,
-    *_hx_utility.__all__,
+    *_heat_exchanging.__all__,
+    *_hydroprocessing.__all__,
+    *_hydrothermal.__all__,
     *_membrane_distillation.__all__,
-    *_phase_changer.__all__,
-    *_pump.__all__,
+    *_pumping.__all__,
     *_reactor.__all__,
-    *_sludge_centrifuge.__all__,
-    *_storage_tank.__all__,
-    *_struvite_precipitation.__all__,
+    *_sludge_thickening.__all__,
+    *_tanks.__all__,
     )
